# NeuronAgent/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: NeuronAgent API
  description: |
    REST API for NeuronAgent - AI agent runtime system providing agent state management,
    session handling, message processing, tool execution, and streaming capabilities.
  version: 1.0.0
  contact:
    name: NeuronDB Support
    email: support@neurondb.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.neurondb.ai/v1
    description: Production server
...
paths:
...
  /agents/{agent_id}/specialization:
    post:
      tags:
        - Specializations
      summary: Create agent specialization
      operationId: createSpecialization
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                specialization_type:
                  type: string
                  enum: [planning, research, coding, execution, analysis, general]
                capabilities:
                  type: array
                  items:
                    type: string
                config:
                  type: object
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialization'

    get:
      tags:
        - Specializations
      summary: Get agent specialization
      operationId: getSpecialization
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialization'

    put:
      tags:
        - Specializations
      summary: Update agent specialization
      operationId: updateSpecialization
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Specialization'

    delete:
      tags:
        - Specializations
      summary: Delete agent specialization
      operationId: deleteSpecialization
      parameters:
        - name: agent_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Specialization deleted
  /specializations:
    get:
      tags:
        - Specializations
      summary: List specializations
      operationId: listSpecializations
      parameters:
        - name: specialization_type
          in: query
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Specialization'
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: API key authentication using Bearer token
...
security:
  - BearerAuth: []